// src/main/menu/context-menu/types/actions.ts
export type ContextMenuActions = {
  showEmojiPanel: () => void;
  openInNewTab: () => void;
  copyLinkURL: () => void;
  toggleNavigationDevTools: () => void;
  toggleDevTools: () => void;
  openSettings: () => void;
  searchSelectionText: () => void;
  startPip: () => void;
  goBack: () => void;
  goForward: () => void;
  reloadTab: () => void;
  viewSource: () => void;
};